Major updates
A controller folder was created with an example posts.js controller.

VerifyToken.js was moved to a middleware folder.

Package.json was edited to allow the use of @folder for in require functions.

Private and Public folders were created in routes to keep protected API routes separate.

all.js was created within /routes to tidy up index.js. All routes are defined in here.